% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_nlmixr.R
\name{assign_origData}
\alias{assign_origData}
\title{Replace the fit data with the original data, then return the modified fit}
\usage{
assign_origData(fit, data)
}
\arguments{
\item{fit}{an estimated \code{nlmixr2} object}

\item{data}{the data from the original fit}
}
\value{
The fit with the data added back in as \code{fit$env$origData}
}
\description{
This function is intended for use within \code{nlmixr2targets} target creation,
and it's not typically invoked by users.
}
\keyword{Internal}
