% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirection.R
\name{nlmixr2_indirect}
\alias{nlmixr2_indirect}
\title{Estimate an nlmixr2 model loading the model from a targets indirect hash storage}
\usage{
nlmixr2_indirect(object, data, est, control)
}
\arguments{
\item{object}{Fitted object or function specifying the model.}

\item{data}{nlmixr data}

\item{est}{estimation method (all methods are shown by
`nlmixr2AllEst()`). Methods can be added for other tools}

\item{control}{The estimation control object.  These are expected
to be different for each type of estimation method}
}
\description{
This is not intended for direct use by users
}
