% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{nlmixr_data_simplify}
\alias{nlmixr_data_simplify}
\title{Standardize and simplify data for nlmixr2 estimation}
\usage{
nlmixr_data_simplify(data, object)
}
\arguments{
\item{data}{nlmixr data}

\item{object}{an nlmixr_ui object (e.g. the output of running
\code{nlmixr(object = model)}}
}
\value{
The data with the nlmixr2 column lower case and on the left and the
covariate columns on the right and alphabetically sorted.
}
\description{
This function is typically not needed by end users.
}
\details{
The standardization keeps columns that rxode2 and nlmixr2 use along with the
covariates.  Column order is standardized (rxode2 then nlmixr2 then
alphabetically sorted covariates), and rxode2 and nlmixr2 column names are
converted to lower case.
}
\seealso{
Other Simplifiers: 
\code{\link{nlmixr_object_simplify}()}
}
\concept{Simplifiers}
