% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{nlmixr_object_simplify}
\alias{nlmixr_object_simplify}
\title{Simplify an nlmixr object}
\usage{
nlmixr_object_simplify(object)
}
\arguments{
\item{object}{Fitted object or function specifying the model.}
}
\value{
\code{object} converted to a nlmixrui object.  The model name is
always "object".
}
\description{
This function is typically not needed by end users.
}
\details{
The object simplification removes comments (so please use \code{label()}
instead of comments to label parameters) and then converts the \code{object}
to a "nlmixrui" object.
}
\seealso{
Other Simplifiers: 
\code{\link{nlmixr_data_simplify}()}
}
\concept{Simplifiers}
