% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_nlmixr_multimodel.R
\name{tar_nlmixr_multimodel_parse}
\alias{tar_nlmixr_multimodel_parse}
\title{Generate nlmixr multimodel target set for all models in one call to
\code{tar_nlmixr_multimodel()}}
\usage{
tar_nlmixr_multimodel_parse(name, data, est, control, table, model_list, env)
}
\arguments{
\item{name}{Symbol, name of the target.
In \code{\link[targets:tar_target]{tar_target()}}, \code{name} is an unevaluated symbol, e.g.
\code{tar_target(name = data)}.
In \code{\link[targets:tar_target_raw]{tar_target_raw()}}, \code{name} is a character string, e.g.
\code{tar_target_raw(name = "data")}.

A target name must be a valid name for a symbol in R, and it
must not start with a dot. Subsequent targets
can refer to this name symbolically to induce a dependency relationship:
e.g. \code{tar_target(downstream_target, f(upstream_target))} is a
target named \code{downstream_target} which depends on a target
\code{upstream_target} and a function \code{f()}. In addition, a target's
name determines its random number generator seed. In this way,
each target runs with a reproducible seed so someone else
running the same pipeline should get the same results,
and no two targets in the same pipeline share the same seed.
(Even dynamic branches have different names and thus different seeds.)
You can recover the seed of a completed target
with \code{tar_meta(your_target, seed)} and run \code{\link[targets:tar_seed_set]{tar_seed_set()}}
on the result to locally recreate the target's initial RNG state.}

\item{data}{nlmixr data}

\item{est}{estimation method (all methods are shown by
`nlmixr2AllEst()`). Methods can be added for other tools}

\item{control}{The estimation control object.  These are expected
to be different for each type of estimation method}

\item{table}{The output table control object (like
`tableControl()`)}

\item{model_list}{A named list of calls for model targets to be created}

\item{env}{The environment where the model is setup (not needed for typical
use)}
}
\description{
Generate nlmixr multimodel target set for all models in one call to
\code{tar_nlmixr_multimodel()}
}
\keyword{Internal}
